requirements.txt
pandas
numpy
scikit-learn
tensorflow
matplotlib
tqdm
pyyaml